requirements.txt
requests
beautifulsoup4
pypdf
python-docx
openpyxl
pandas
openai
